moved setup file to config folder with a similar file that has eslint disable statement as well
