+++ tests/setup/configureSmtp.spec.ts
@@ -1,7 +1,7 @@
 import { it, expect, vi, describe } from "vitest";
 import inquirer from "inquirer";
 import * as verifySmtpConnection from "../../src/setup/verifySmtpConnection";
-import { configureSmtp } from "../../setup";
+import { configureSmtp } from "../../config/setup";
 import dotenv from "dotenv";
 import fs from "fs";
